LPPaver.Constraint.Type

Description

Module defining the Constraint data type.

Documentation

data Constraint

The Constraint data type is used to represent non-strict inequalities. The LHS is a list of String variables along with their Rational coefficients. The RHS is a Rational number.

Constructors

GEQ [(String, Rational)] Rational

LEQ [(String, Rational)] Rational