LPPaver.Constraint.Util

Description

Module defining useful utility functions for the Constraint data type.

Documentation

constraintLeftSide :: Constraint -> [(String, Rational)]

Get the LHS of a Constraint

constraintRightSide :: Constraint -> Rational

Get the LHS of a Constraint

constraintVars :: [Constraint] -> [String]

Get a list of all variables in a list of Constraints

constraintsToSimplexConstraints :: [Constraint] -> ([PolyConstraint], Map String Integer)

Convert a list of Constraints into a list of PolyConstraints. The function returns a pair with the first item being a list of PolyConstraints and the second item being a mapping of String variables (from Constraints) to Integer variables (for PolyConstraints).